requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
bcrypt
pydantic
email-validator